Rearranged test files in test directories in php-wasm
