+++ packages/php-wasm/universal/src/test/process-id-allocator.spec.ts
@@ -1,5 +1,4 @@
-import { describe, it, expect } from 'vitest';
-import { ProcessIdAllocator } from './process-id-allocator';
+import { ProcessIdAllocator } from '../lib/process-id-allocator';
 
 describe('ProcessIdAllocator', () => {
 	it('claims IDs starting from initialId', () => {
