+++ packages/php-wasm/web-service-worker/src/lib/fetch-with-cors-proxy.ts
