+++ packages/php-wasm/web-service-worker/src/lib/firewall-interference-error.ts
