+++ packages/php-wasm/web-service-worker/src/lib/messaging.ts
