+++ packages/php-wasm/web-service-worker/src/lib/utils.ts
